Split app in pages
