+++ src/pages/Items.js
@@ -1,0 +1,7 @@
+import React from "react";
+
+function Items(props) {
+  return <div>ITEMS</div>;
+}
+
+export default Items;
